requirements.txt
asgiref==3.3.1
certifi==2021.5.30
charset-normalizer==2.0.4
defusedxml==0.7.1
diff-match-patch==20200713
dj-database-url==0.5.0
Django==3.1.6
django-crispy-forms==1.12.0
django-email-verification==0.1.0
django-import-export==2.5.0
et-xmlfile==1.1.0
gunicorn==20.1.0
heroku==0.1.4
idna==3.2
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.0.7
Pillow==8.3.1
psycopg2==2.8.6
python-dateutil==1.5
pytz==2021.1
PyYAML==5.4.1
requests==2.26.0
sqlparse==0.4.1
tablib==3.0.0
urllib3==1.26.6
whitenoise==5.3.0
xlrd==2.0.1
xlwt==1.3.0
